+++ services/ifsv/ifd/ifd_api.c
@@ -1,18 +1,18 @@
 /*      -*- OpenSAF  -*-
  *
- * (C) Copyright 2008 The OpenSAF Foundation 
+ * (C) Copyright 2008 The OpenSAF Foundation
  *
  * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY 
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
  * or FITNESS FOR A PARTICULAR PURPOSE. This file and program are licensed
  * under the GNU Lesser General Public License Version 2.1, February 1999.
  * The complete license can be accessed from the following location:
- * http://opensource.org/licenses/lgpl-license.php 
+ * http://opensource.org/licenses/lgpl-license.php
  * See the Copying file included with the OpenSAF distribution for full
  * licensing terms.
  *
  * Author(s): Emerson Network Power
- *   
+ *
  */
 
 
@@ -313,7 +313,7 @@ ifd_lib_init (IFSV_CREATE_PWE *pwe_param)
       {
          printf("m_NCS_TASK_START  failed\n");
          m_IFD_LOG_SYS_CALL_FAIL(IFSV_LOG_TASK_CREATE_FAIL,\
-         m_IFSV_GET_TASK_HDL(vrid,comp_type));
+         (long)m_IFSV_GET_TASK_HDL(vrid,comp_type));
          goto ifd_task_start_fail;
       }
