+++ services/ifsv/ifd/ifd_saf.c
@@ -1,18 +1,18 @@
 /*      -*- OpenSAF  -*-
  *
- * (C) Copyright 2008 The OpenSAF Foundation 
+ * (C) Copyright 2008 The OpenSAF Foundation
  *
  * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY 
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
  * or FITNESS FOR A PARTICULAR PURPOSE. This file and program are licensed
  * under the GNU Lesser General Public License Version 2.1, February 1999.
  * The complete license can be accessed from the following location:
- * http://opensource.org/licenses/lgpl-license.php 
+ * http://opensource.org/licenses/lgpl-license.php
  * See the Copying file included with the OpenSAF distribution for full
  * licensing terms.
  *
  * Author(s): Emerson Network Power
- *   
+ *
  */
 
 
@@ -390,7 +390,7 @@ ifd_saf_health_chk_callback (SaInvocationT invocation,
       saAmfResponse(ifsv_cb->amf_hdl,invocation, error);
       m_IFSV_GIVE_CB(ifsv_cb);
    }  
-   m_IFD_LOG_API_L(IFSV_LOG_AMF_HEALTH_CHK,checkType->key); 
+   m_IFD_LOG_API_L(IFSV_LOG_AMF_HEALTH_CHK,(long)checkType->key); 
    return;
 }
 /****************************************************************************
