+++ services/ifsv/ifnd/ifnd_mds.c
@@ -1,18 +1,18 @@
 /*      -*- OpenSAF  -*-
  *
- * (C) Copyright 2008 The OpenSAF Foundation 
+ * (C) Copyright 2008 The OpenSAF Foundation
  *
  * This program is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY 
+ * WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
  * or FITNESS FOR A PARTICULAR PURPOSE. This file and program are licensed
  * under the GNU Lesser General Public License Version 2.1, February 1999.
  * The complete license can be accessed from the following location:
- * http://opensource.org/licenses/lgpl-license.php 
+ * http://opensource.org/licenses/lgpl-license.php
  * See the Copying file included with the OpenSAF distribution for full
  * licensing terms.
  *
  * Author(s): Emerson Network Power
- *   
+ *
  */
 
 
@@ -524,7 +524,7 @@ static uns32 ifnd_mds_rcv(IFSV_CB *cb, MDS_CALLBACK_RECEIVE_INFO *rcv_info)
   {
        m_IFND_LOG_STR_2_NORMAL(IFSV_LOG_FUNC_RET_FAIL,
              "IFND::m_IFND_EVT_SEND  FAILED ","FUNC::ifnd_mds_rcv");
-     m_IFND_LOG_SYS_CALL_FAIL(IFSV_LOG_MSG_QUE_SEND_FAIL,mbx);
+     m_IFND_LOG_SYS_CALL_FAIL(IFSV_LOG_MSG_QUE_SEND_FAIL,(long)mbx);
      return (NCSCC_RC_FAILURE);
   }
